backend/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
neo4j==5.28.1
numpy==2.2.6
paho-mqtt==2.1.0
pandas==2.2.3
psycopg2-binary==2.9.10
PyJWT==2.9.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
pymysql
gunicorn
cryptography
redis
requests
